nerf/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ns-nerf"
version = "0.1.0"
description = "Video-to-Nerf (Nerfstudio) pipeline with clean, reproducible layout."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Xinjie" }]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13",
]

[project.scripts]
ns-nerf = "ns_nerf.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "ns_nerf",
    "ns_nerf.pipeline",
]
